# specs/001-meeting-transcription-mvp/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: RAG-Enhanced Meeting Intelligence System API
  description: |
    FastAPI REST API for the RAG-Enhanced Meeting Intelligence System.

    This system extends existing Google Meet transcription infrastructure with RAG capabilities,
    supporting multi-source file ingestion (meeting transcripts, PDFs, DOCX, JSON, markdown),
    dual-provider LLM integration (OpenAI, Anthropic) with user-selectable configuration, declarative artifact generation,
    and intelligent RAG chat with citations.

    **Key Features**:
    - File upload and ingestion with automatic processing
    - Job tracking with detailed step-by-step logs
    - Artifact querying (summaries, decisions, entities, timelines)
    - RAG chat with speaker-aware citations for meetings
    - Backfill operations for historical data processing
    - Health checks and system status
  version: 1.0.0
  contact:
    name: API Support
    email: support@meethub.local
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://localhost:8000/api/v1
    description: API v1 base path
security:
  - oauth2: []
  - apiKey: []
...
paths:
...
  /api/v1/files/upload:
    post:
      summary: Upload file for ingestion
      description: |
        Upload a file for RAG ingestion. Supports multiple file types: markdown, PDF, DOCX, JSON, plain text.
        Maps to FR-008 (calculate SHA-256 hash), FR-009 (create File and FileVersion entities).

        **User Stories**: US1 (Manual File Ingestion), US6 (Backfill)
        **Functional Requirements**: FR-008, FR-009, FR-010, FR-011
      tags: [Files]
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '202':
          description: File queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
              example:
                file_id: 123
                file_version_id: 456
                job_run_id: 789
                status: "queued"
                message: "File queued for ingestion"
        '400':
          description: Invalid file or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                unsupportedFileType:
                  value:
                    error: "UnsupportedFileType"
                    message: "File type .exe is not supported"
                    details: { "supported_types": ["md", "pdf", "docx", "json", "txt"] }
                fileTooBig:
                  value:
                    error: "FileTooLarge"
                    message: "File size exceeds 100MB limit"
                    details: { "file_size_mb": 125.4, "max_size_mb": 100 }
        '409':
          description: Duplicate file (identical SHA-256 hash already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "DuplicateFile"
                message: "File with identical content already exists"
                details:
                  existing_file_id: 123
                  existing_file_version_id: 456
                  content_hash: "a1b2c3..."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/files:
    get:
      summary: List files with filtering
      description: |
        List ingested files with optional filters.
        Maps to data management and file tracking (FR-070 to FR-074).

        **Functional Requirements**: FR-070 (soft delete), FR-071 (version history)
      tags: [Files]
      operationId: listFiles
      parameters:
        - $ref: '#/components/parameters/ProjectIdQuery'
        - name: source_type
          in: query
          schema:
            $ref: '#/components/schemas/FileSourceType'
        - name: mime_type
          in: query
          schema:
            type: string
            example: "application/pdf"
        - name: deleted
          in: query
          description: Include deleted files
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: File list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/files/{file_id}:
    get:
      summary: Get file details with version history
      description: |
        Retrieve file details including all versions and associated artifacts.
        Maps to FR-071 (maintain file version history).

        **User Stories**: US7 (Change Detection & Incremental Updates)
      tags: [Files]
      operationId: getFileById
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      summary: Soft delete file
      description: |
        Mark file as deleted (soft delete). File and versions retained but excluded from queries.
        Maps to FR-070 (implement soft delete).

        **Functional Requirements**: FR-070
      tags: [Files]
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: hard_delete
          in: query
          description: Permanently delete file and all related data (use with caution)
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: File deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "File soft deleted successfully"
                  file_id:
                    type: integer
                  deleted:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/files/{file_id}/versions:
    get:
      summary: Get file version history
      description: |
        Retrieve all versions of a file with hashes and timestamps.
        Maps to FR-071 (maintain file version history).
      tags: [Files]
      operationId: getFileVersions
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: integer
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/FileVersion'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/files/{file_id}/reprocess:
    post:
      summary: Trigger file reprocessing
      description: |
        Force reprocessing of file through RAG pipeline (re-chunk, re-embed, regenerate artifacts).
        Useful after rule changes or to retry failed processing.

        **User Stories**: US7 (Incremental Updates)
      tags: [Files]
      operationId: reprocessFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '202':
          description: Reprocessing queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_run_id:
                    type: integer
                  status:
                    type: string
                    example: "queued"
        '404':
          $ref: '#/components/responses/NotFound'

  # ============================================================================
  # JOB TRACKING ENDPOINTS (FR-043 to FR-049)
  # ============================================================================
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Google OAuth2 authentication (existing google_auth.py flow)
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect
            email: User email address
            profile: User profile information

    apiKey:
      type: apiKey
      description: API key authentication (future extension)
      name: X-API-Key
      in: header
  parameters:
...
    ProjectIdQuery:
      name: project_id
      in: query
      schema:
        type: integer
      description: Filter by project ID
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
      description: Maximum number of results to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of results to skip (pagination)
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Valid authentication credentials required"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "InternalServerError"
            message: "An unexpected error occurred"
            details:
              traceback: "..."
  schemas:
...
    FileSourceType:
      type: string
      enum:
        - meeting
        - uploaded_document
        - chat_export
      description: |
        Source classification for files:
        - `meeting`: File created from meeting transcript bridge
        - `uploaded_document`: Manually uploaded file
        - `chat_export`: Chat backup (Slack, Teams, WhatsApp)
...
    FileUploadRequest:
      type: object
      required:
        - file
      properties:
        file:
          type: string
          format: binary
          description: File to upload (max 100MB)
        project_id:
          type: integer
          description: Target project ID (uses default if not specified)
        force_reprocess:
          type: boolean
          default: false
          description: Force reprocessing even if identical hash exists
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code/type
          example: "ValidationError"
        message:
          type: string
          description: Human-readable error message
          example: "File type not supported"
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    FileUploadResponse:
      type: object
      required:
        - file_id
        - file_version_id
        - status
      properties:
        file_id:
          type: integer
        file_version_id:
          type: integer
        job_run_id:
          type: integer
          description: Job ID for tracking processing
        status:
          type: string
          enum: [queued, duplicate_skipped]
        message:
          type: string
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...